Improve README.md
The README.md should be a brief abstract of information interesting to
users. This commit improves the project description as well as the
usage section while it replaces the design section with an installation
section as the latter is arguably a more useful information for users
of an application. Furthermore, the commit fixes the link to the
license.