Refine documentation and testing policy
The current documentation policy is somewhat abstract. Here we try to
make it more concrete while retaining the flexibilty of the old one in
a mostly backwards-compatible way. Most importantly, we specify how to
document classes and functions. The hope is also to simplify the
documentation process, especially for new and occasional contributers.

The testing policy is enhanced with a guideline of what to test with a
reasoning similar to that of the documentation policy change.